Keep existing values on `cm init`
It turns out `cm init` has access to the existing parsed config so we
can use those in templates (if they exist) to avoid replacing all values
with defaults on `cm init`.

After trying a few variations, ended up using `with` to scope-down the
pipeline, and `pluck` to get the existing value, or default. This works
better than `default`, as we want to know whether the existing value
exists, rather than whether it's the default (we don't want to replace
with a default if it's set).

`$defaultDict` is necessary since `with` will skip empty pipelines by
default, and we want to ensure that we bootstrap the default values for
each field in the config.